pyproject.toml
[project]
name = "anews"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "feedparser",
    "httpx",
    "psycopg2-binary",
    "python-dotenv",
    "beautifulsoup4",
    "trafilatura>=2.0.0",
    "sentence-transformers",
    "openai-whisper>=20250625",
    "whisperx>=3.8.2",
    "pyannote-audio>=4.0.4",
    "alembic>=1.18.4",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
